PyTorch/Recommendation/DLRM/requirements.txt
git+https://github.com/NVIDIA/dllogger#egg=dllogger
pandas
absl-py>=0.7.0
